+++ server/apps/trak/models/__init__.py
@@ -2,7 +2,8 @@ from .address import Address
 from .contact import Contact, EpaPhone
 from .handler import Handler
 from .manifest import Manifest
-from .rcra_profile import RcraProfile, SitePermission
+from .rcra_profile import RcraProfile
+from .site_permission import SitePermission
 from .sites import Site
 from .transporter import Transporter
 from .waste import WasteLine
